+++ src/ImpromptuModular.hpp
@@ -43,6 +43,7 @@ extern Model *modelPhraseSeq32;
 extern Model *modelPhraseSeqExpander;
 extern Model *modelProbKey;
 extern Model *modelSemiModularSynth;
+extern Model *modelSygen;
 extern Model *modelTact;
 extern Model *modelTact1;
 extern Model *modelTactG;